moved tests to proper packages
