+++ src/test/java/com/solab/iso8583/TestIssue4.java
@@ -1,7 +1,5 @@
-package j8583;
+package com.solab.iso8583;
 
-import com.solab.iso8583.IsoMessage;
-import com.solab.iso8583.MessageFactory;
 import com.solab.iso8583.parse.ConfigParser;
 import org.junit.Assert;
 import org.junit.Test;
